tools/devops/requirements.txt
azure-devops==7.1.0b4
click==8.1.3
gitpython==3.1.41
backoff==2.2.1